Extracted out export/import to `FileHandler` class
